+++ src/main/kotlin/org/jetbrains/research/testspark/settings/llm/SettingsLLMComponent.kt
@@ -14,8 +14,8 @@ import com.intellij.util.ui.FormBuilder
 import org.jdesktop.swingx.JXTitledSeparator
 import org.jetbrains.research.testspark.bundles.TestSparkLabelsBundle
 import org.jetbrains.research.testspark.bundles.TestSparkToolTipsBundle
-import org.jetbrains.research.testspark.actions.llm.addLLMPanelListeners
-import org.jetbrains.research.testspark.actions.llm.isGrazieClassLoaded
+import org.jetbrains.research.testspark.helpers.addLLMPanelListeners
+import org.jetbrains.research.testspark.helpers.isGrazieClassLoaded
 import org.jetbrains.research.testspark.services.PromptParserService
 import org.jetbrains.research.testspark.settings.SettingsApplicationState
 import java.awt.FlowLayout
