+++ src/main/kotlin/org/jetbrains/research/testspark/tools/Pipeline.kt
@@ -8,10 +8,10 @@ import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.progress.Task
 import com.intellij.openapi.roots.ProjectFileIndex
 import com.intellij.openapi.roots.ProjectRootManager
-import org.jetbrains.research.testspark.actions.getSurroundingClass
 import org.jetbrains.research.testspark.bundles.TestSparkBundle
 import org.jetbrains.research.testspark.data.DataFilesUtil
 import org.jetbrains.research.testspark.data.FragmentToTestData
+import org.jetbrains.research.testspark.helpers.getSurroundingClass
 import org.jetbrains.research.testspark.services.ClearService
 import org.jetbrains.research.testspark.services.ProjectContextService
 import org.jetbrains.research.testspark.services.TestStorageProcessingService
