+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/Patterns.kt
@@ -1,0 +1,16 @@
+package org.jetbrains.research.testspark.tools.llm.generation
+
+val importPattern = Regex(
+    pattern = "^import\\s+(static\\s)?((?:[a-zA-Z_]\\w*\\.)*[a-zA-Z_](?:\\w*\\.?)*)(?:\\.\\*)?;",
+    options = setOf(RegexOption.MULTILINE),
+)
+
+val packagePattern = Regex(
+    pattern = "^package\\s+((?:[a-zA-Z_]\\w*\\.)*[a-zA-Z_](?:\\w*\\.?)*)(?:\\.\\*)?;",
+    options = setOf(RegexOption.MULTILINE),
+)
+
+val runWithPattern = Regex(
+    pattern = "@RunWith\\([^)]*\\)",
+    options = setOf(RegexOption.MULTILINE),
+)
